More feedback from thinky puzzle games group
Reset input methods after resetting or (un)pausing the game.
Buttons should now be in the same positions for all input methods (AZERTY and DVORAK use the same key positions as QWERTY).
Changed how targets & walls look. Targets gain alpha when a block is placed on it (or the player in the level select)